consolidate files for sim model and state
