+++ sendnn_inference/v1/sim.py
@@ -1,23 +1,26 @@
-"""Process-local virtual-clock state for sim mode.
+"""Sim-mode plumbing: no-op model + virtual-clock state.
 
-The runner advances the clock once per forward step (charging prefill_ms or
-decode_ms) and per-request bookkeeping accumulates in `_records`. When a
-request finishes, the runner calls `finalize_and_write` which appends a
-JSONL line of virtual stats to `<perf_dir>/sim_metrics.jsonl`.
+Activated by SENDNN_INFERENCE_SIM_MODE=1. The runner instantiates
+``MockSpyreCausalLM`` instead of ``SpyreCausalLM`` (no FMS load, no
+torch.compile, no senlib) and feeds each forward step into ``SimState``,
+which advances a virtual clock by ``SIM_PREFILL_MS`` or ``SIM_DECODE_MS``
+and accumulates per-request timing. When a request finishes, the runner
+calls ``finalize_and_write`` which appends a JSONL line of virtual stats
+to ``<perf_dir>/sim_metrics.jsonl``.
 
-A separate file (rather than substituting into vLLM's request_metrics.jsonl)
-avoids the AsyncLLM process boundary: the FileStatLogger that emits
-request_metrics.jsonl runs in a different process from the runner, so it
-cannot see SimState. Sim mode disables that logger so only sim_metrics.jsonl
-is written.
+A separate output file (rather than substituting into vLLM's
+request_metrics.jsonl) avoids the AsyncLLM process boundary: the
+FileStatLogger that emits request_metrics.jsonl runs in a different
+process from the runner, so it cannot see SimState. Sim mode disables
+that logger so only sim_metrics.jsonl is written.
 
 Token timestamps and ITL: each forward step advances the global virtual
-clock. We record, per request, the end-time of every prefill step and every
-decode step it participates in. The first sampled token is produced by the
-*last* prefill chunk; subsequent tokens come from each decode step. This
-gives a per-token virtual timeline and meaningful ITL — the gap between
-two consecutive decode tokens widens whenever an intervening prefill of
-another request happens.
+clock. We record, per request, the end-time of every prefill step and
+every decode step it participates in. The first sampled token is produced
+by the *last* prefill chunk; subsequent tokens come from each decode
+step. This gives a per-token virtual timeline and a meaningful ITL — the
+gap between two consecutive decode tokens widens whenever an intervening
+prefill of another request happens.
 """
 
 import json
@@ -25,10 +28,107 @@ from dataclasses import dataclass, field
 from datetime import datetime
 from pathlib import Path
 from threading import Lock
+from types import SimpleNamespace
 
+import torch
+from vllm.config import VllmConfig
+from vllm.forward_context import get_forward_context
 from vllm.v1.core.sched.output import SchedulerOutput
+from vllm.v1.outputs import SamplerOutput
+from vllm.v1.sample.metadata import SamplingMetadata
+from vllm.v1.sample.sampler import Sampler
 
 import sendnn_inference.envs as envs_spyre
+from sendnn_inference.model_executor.model_loader.spyre import SpyreAttentionMetadata
+
+
+# ---------------------------------------------------------------------------
+# Mock model
+# ---------------------------------------------------------------------------
+
+
+class MockSpyreCausalLM:
+    """No-op stand-in for SpyreCausalLM.
+
+    Returns dummy logits without running any real forward pass. Also used
+    by unit tests that exercise scheduler/runner logic without a real model.
+    """
+
+    def __init__(
+        self,
+        vllm_config: VllmConfig,
+    ) -> None:
+        self.sampler = Sampler()
+
+        # boolean tensor of length batch size with indices:
+        # True for unfinished sequences and
+        # False for finished or padded sequences
+        self.indices = None
+
+        # number of right pads (relevant for continuous batching only)
+        self.n_pads_right = 0
+
+        self.vocab_size = vllm_config.model_config.get_vocab_size()
+
+        # ChunkedPrefillModelRunner.vocab_size reads .fms_model.config.src_vocab_size
+        # and .is_multimodal directly; provide minimal shims so warmup works.
+        self.is_multimodal = False
+        self.fms_model = SimpleNamespace(config=SimpleNamespace(src_vocab_size=self.vocab_size))
+
+        # These variables are here for future test scenarios to use
+        self.last_input_ids: torch.Tensor | None = None
+        self.last_positions: torch.Tensor | None = None
+        self.last_masks: torch.Tensor | None = None
+        self.last_is_prompt: bool | None = None
+        self.last_attn_metadata: SpyreAttentionMetadata | None = None
+
+    def get_maybe_mm_embeddings(self, *args, **kwargs):
+        # This model is not multimodal
+        return None
+
+    def __call__(self, *args, **kwargs):
+        return self.forward(*args, **kwargs)
+
+    def forward(
+        self,
+        input_ids_or_embeds: torch.Tensor,
+        positions: torch.Tensor,
+        masks: torch.Tensor,
+        is_prompt: bool,
+    ) -> torch.Tensor:
+        # These variables are here for future test scenarios to use;
+        # NOTE: for now, we always use input IDs since this isn't multimodal.
+        self.last_input_ids = input_ids_or_embeds
+        self.last_positions = positions
+        self.last_masks = masks
+        self.last_is_prompt = is_prompt
+
+        forward_context = get_forward_context()
+
+        assert isinstance(forward_context.attn_metadata, SpyreAttentionMetadata)
+        self.last_attn_metadata = forward_context.attn_metadata
+
+        batch_size = input_ids_or_embeds.shape[0]
+
+        return torch.empty(
+            (batch_size, self.vocab_size), dtype=torch.float32, device=input_ids_or_embeds.device
+        )
+
+    def sample(
+        self,
+        logits: torch.Tensor,
+        sampling_metadata: SamplingMetadata,
+    ) -> SamplerOutput | None:
+        next_tokens = self.sampler(logits, sampling_metadata)
+        return next_tokens
+
+    def set_past_key_value_states(self, num_blocks) -> None:
+        pass
+
+
+# ---------------------------------------------------------------------------
+# Virtual-clock state
+# ---------------------------------------------------------------------------
 
 
 @dataclass
